pyproject.toml
[project]
name = "humchess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-chess>=1.999",
    "torch>=2.9.1",
    "PyYAML>=6.0.2",
    "zstandard>=0.25.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
